# packages/contracts/openapi.json
{
  "openapi": "3.1.0",
  "paths": {
...
    "/staff/users/{id}/suspend": {
      "post": {
        "operationId": "StaffUserController_suspend",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SuspendUserDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "staff-users"
        ]
      }
    },
    "/staff/users/{id}/block": {
      "post": {
        "operationId": "StaffUserController_block",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BlockUserDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "staff-users"
        ]
      }
    },
    "/staff/users/{id}/reactivate": {
      "post": {
        "operationId": "StaffUserController_reactivate",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "staff-users"
        ]
      }
    },
...
  },
  "info": {
    "title": "Bobo&Doda API",
    "description": "Markaziy Osiyo freelance marketplace — milestone escrow, double-entry ledger. Frontend `lib/api` chegarasini qoplaydi.",
    "version": "0.1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "SuspendUserDto": {
        "type": "object",
        "properties": {
          "reason": {
            "type": "string",
            "minLength": 5,
            "maxLength": 1000
          },
          "suspendedUntil": {
            "type": "string",
            "description": "ISO sana — berilmasa muddatsiz (staff qo‘lda tugatadi)"
          }
        },
        "required": [
          "reason"
        ]
      },
      "BlockUserDto": {
        "type": "object",
        "properties": {
          "reason": {
            "type": "string",
            "minLength": 5,
            "maxLength": 1000
          }
        },
        "required": [
          "reason"
        ]
      },
...
    }
  }
}